Compiling with MSVC.
1. std::min -> (std::min), as windows.h defines preprocessor macro min(a,b).
(Another way to fix this was: #define NOMINMAX before each #include <windows.h>)
2. unsigned int -> DWORD in the last argument to RegQueryValueEx.
3. Forbid #include <errno.h> in MSVC, as there is no errno.h for WM 2005 Smartphome target in MSVC 2005.

Didn't check this under Linux, hope didn't break anything.

+++ src/Track.cpp
@@ -241,7 +241,7 @@ void CTrack::FlushGPX(int iSize)
 		{
 			DWORD fileSize = GetFileSize(hFile, NULL);
 			if (fileSize >= iCloseLength)
-				SetFilePointer(hFile, -iCloseLength, 0, FILE_END);
+				SetFilePointer(hFile, -LONG(iCloseLength), 0, FILE_END);
 			DWORD dwWritten;
 			WriteFile(hFile, m_writeBuffer, iSize, &dwWritten, NULL);
 			WriteFile(hFile, sCloseXml, iCloseLength, &dwWritten, NULL);
